benchmark: split private functions to separate files
