+++ lxd-benchmark/main.go
@@ -15,6 +15,7 @@ var argCount = gnuflag.Int("count", 100, "Number of containers to create")
 var argParallel = gnuflag.Int("parallel", -1, "Number of threads to use")
 var argImage = gnuflag.String("image", "ubuntu:", "Image to use for the test")
 var argPrivileged = gnuflag.Bool("privileged", false, "Use privileged containers")
+var argStart = gnuflag.Bool("start", true, "Start the container after creation")
 var argFreeze = gnuflag.Bool("freeze", false, "Freeze the container right after start")
 
 func main() {
@@ -41,7 +42,7 @@ func run(args []string) error {
 		}
 		gnuflag.SetOut(out)
 
-		fmt.Fprintf(out, "Usage: %s spawn [--count=COUNT] [--image=IMAGE] [--privileged=BOOL] [--parallel=COUNT]\n", os.Args[0])
+		fmt.Fprintf(out, "Usage: %s spawn [--count=COUNT] [--image=IMAGE] [--privileged=BOOL] [--start=BOOL] [--freeze=BOOL] [--parallel=COUNT]\n", os.Args[0])
 		fmt.Fprintf(out, "       %s delete [--parallel=COUNT]\n\n", os.Args[0])
 		gnuflag.PrintDefaults()
 		fmt.Fprintf(out, "\n")
@@ -50,7 +51,7 @@ func run(args []string) error {
 			return nil
 		}
 
-		return fmt.Errorf("An valid action (spawn or delete) must be passed.")
+		return fmt.Errorf("A valid action (spawn or delete) must be passed.")
 	}
 
 	gnuflag.Parse(true)
@@ -65,7 +66,7 @@ func run(args []string) error {
 
 	switch os.Args[1] {
 	case "spawn":
-		_, err = benchmark.SpawnContainers(c, *argCount, *argParallel, *argImage, *argPrivileged, *argFreeze)
+		_, err = benchmark.SpawnContainers(c, *argCount, *argParallel, *argImage, *argPrivileged, *argStart, *argFreeze)
 		return err
 	case "delete":
 		containers, err := benchmark.GetContainers(c)
